use Pos, move to separate file
